Create tritjs-2.0.ebuild
Below is the updated tritjs-1.0.ebuild with added documentation in the form of inline comments and a README file installed as part of the package. These instructions will guide users on how to use the package, including enabling the tri USE flag, running the binary, and understanding the build process. I’ve also included a simple README generated during src_prepare() to ensure documentation is always present, even if the doc USE flag is disabled.